Refactor/data: data 폴더 내부 code 정리 & 폴더명 변경
