+++ app/src/main/java/com/depromeet/housekeeper/data/repository/RemoteRepository.kt
@@ -1,6 +1,6 @@
 package com.depromeet.housekeeper.data.repository
 
-import com.depromeet.housekeeper.data.remote.RemoteDataSource
+import com.depromeet.housekeeper.data.dataSource.RemoteDataSource
 import com.depromeet.housekeeper.model.HouseWork
 import com.depromeet.housekeeper.model.request.*
 import com.depromeet.housekeeper.model.response.*
