+++ app/src/main/java/com/depromeet/housekeeper/data/repository/Repository.kt
@@ -1,7 +1,7 @@
 package com.depromeet.housekeeper.data.repository
 
 import com.depromeet.housekeeper.data.RetrofitBuilder
-import com.depromeet.housekeeper.data.remote.RemoteDataSource
+import com.depromeet.housekeeper.data.dataSource.RemoteDataSource
 import com.depromeet.housekeeper.model.HouseWork
 import com.depromeet.housekeeper.model.request.UpdateChoreBody
 import com.depromeet.housekeeper.model.request.UpdateMember
